move details thunk to thunk file, move tests, clean up with details hoc
