+++ src/modules/data/details/actions.js
@@ -1,11 +1,7 @@
-import { isEmpty } from 'lodash';
-
 /**
  * Internal dependencies
  */
 import * as types from './types';
-import * as selectors from './selectors';
-import { actions as requestActions } from 'data/request';
 
 export const enableLoading = ( id ) => ( {
 	type: types.SET_DETAILS_LOADING,
@@ -30,31 +26,6 @@ export const setDetails = ( id, details ) => ( {
 		details,
 	},
 } );
-
-export const fetchDetails = ( id ) => ( dispatch, getState ) => {
-	const state = getState();
-	const props = { name: id };
-	const loading = selectors.getLoading( state, props );
-	const details = selectors.getDetails( state, props );
-
-	if ( ! isEmpty( details ) || loading ) {
-		return;
-	}
-
-	const type = selectors.getPostType( state, props );
-	const options = {
-		path: `${ type }/${ id }`,
-		actions: {
-			start: () => dispatch( enableLoading( id ) ),
-			success: ( { body } ) => {
-				dispatch( setDetails( id, body ) );
-				dispatch( disableLoading( id ) );
-			},
-			error: () => dispatch( disableLoading( id ) ),
-		},
-	};
-	dispatch( requestActions.wpRequest( options ) );
-};
 
 export const setPostType = ( id, type ) => ( {
 	type: types.SET_DETAILS_POST_TYPE,
