+++ src/modules/data/details/index.js
@@ -4,7 +4,8 @@
 import reducer from './reducer';
 import * as types from './types';
 import * as actions from './actions';
+import * as thunks from './thunks';
 import * as selectors from './selectors';
 
 export default reducer;
-export { types, actions, selectors };
+export { types, actions, thunks, selectors };
